requirements.txt
openpyxl
python-docx
matplotlib
sumy
